added touch events
tested on ipad. I turned off the e.preventdefault() because i needed the touch to still come through to select objects in my case. The only other change was using event.layerX to get eventPoint accurately on my svg element which was positioned using css and not at 0,0 like in the demos